refactor(document): split source editing responsibilities
